Los nombres de los paquetes van en minuscula por convención
Los nombres de los paquetes van en minúscula por convención

Es una pavada pero Eclipse me marca el warning
