+++ src/main/java/pexception/FactoryException.java
@@ -1,4 +1,4 @@
-package PException;
+package pexception;
 
 import sun.java2d.loops.CustomComponent;
