+++ src/main/java/pexception/FactoryHabitacionException.java
@@ -1,4 +1,4 @@
-package PException;
+package pexception;
 
 public class FactoryHabitacionException extends  Exception{
     public FactoryHabitacionException(String mensaje){super(mensaje);}
