+++ src/main/java/ppersona/pmedico/Pediatria.java
@@ -1,7 +1,10 @@
-package PPersona.PMedico;
+package ppersona.pmedico;
 
-public class Clinico extends Medico{
-    public Clinico(String nombre,String apellido){
+
+
+public class Pediatria extends Medico{
+
+    public Pediatria(String nombre,String apellido){
         this.nombre=nombre;
         this.apellido=apellido;
         iniciaMatricula();
@@ -67,10 +70,9 @@ public class Clinico extends Medico{
 
 
 
-
     @Override
     public String saludo() {
-        return super.saludo()+" Clinico";
+        return super.saludo() + " Pediatra";
     }
 
     @Override
@@ -78,11 +80,10 @@ public class Clinico extends Medico{
         return this.matricula;
     }
 
-
     @Override
     public float getHonorario() {
 
-        float bonificacion = (float) 1.05;
+        float bonificacion = (float) 1.07;
         return super.getHonorario()* bonificacion;
     }
 }
